Make GUI react only on finger which was down inside GUI element.

+++ src/beryll/core/EventHandler.cpp
@@ -94,7 +94,7 @@ namespace Beryll
                                                  glm::vec2(event.tfinger.x * Window::getInstance()->getScreenWidth(), event.tfinger.y * Window::getInstance()->getScreenHeight()),
                                                   false,
                                                   true,
-                                                  static_cast<int>(event.tfinger.fingerId)});
+                                                  event.tfinger.fingerId});
                     break;
 
                 case SDL_FINGERUP:
